+++ .eslintrc.yml
@@ -1,0 +1,29 @@
+---
+extends: airbnb
+env:
+  node: true
+  es6: true
+globals:
+  config: true
+parserOptions:
+  ecmaVersion: 7
+  ecmaFeatures:
+    jsx: true
+    experimentalObjectRestSpread: true
+  sourceType: module
+plugins: [react]
+
+rules:
+  max-len: [1, 140, 2, {"ignoreUrls":true}]
+  comma-dangle: [error, never]
+  no-console: [error, allow: [warn, error, info]]
+settings:
+  import/resolver:
+    node:
+      extensions:
+        - .js
+        - .jsx
+
+      moduleDirectory:
+        - node_modules
+        - app
